# improve performance / allocations for registrations resolve